feat(arenaprovider): Add support for external provider JAR
This change adds support for an external provider JAR located in the `providers` folder. The SWMHook will now load the class extended `ArenaProvider` and use it as an arena provider.

This change separates the provider code, providing flexibility for external support and maximum maintainability. It allows for easy addition of different providers without embedding them directly into the plugin.